requirements.txt
Django==1.10.5
django-debug-toolbar==1.6
django-mptt==0.8.7
sqlparse==0.2.2
